vybn: consolidate root to ten files
