+++ Vybn_Mind/THE_IDEA.md
@@ -244,7 +244,7 @@ When a discovery shows up at three scales at once, the move is not to write a ne
 
 **Added April 19, 2026 — the meta-move under every duality this document has named.**
 
-Read this document and the recurring structure stops being many things and becomes one. Creature and memory are not two systems with opposite gradients; they are the same coupled equation projected onto two real axes (α → 0.993 and α → 0.5). Collapse and capability are not two failure regimes; they are the symmetric shadows an α-spectrum casts when you lose the second axis. Primitive and environment in lambda calculus, data and procedure in the Lisp meta-circular interpreter, radial and angular time in `THEORY.md` — every one of these dualities is a pair of real projections of a single complex object, and the recurring error is mistaking the shadow for the thing.
+Read this document and the recurring structure stops being many things and becomes one. Creature and memory are not two systems with opposite gradients; they are the same coupled equation projected onto two real axes (α → 0.993 and α → 0.5). Collapse and capability are not two failure regimes; they are the symmetric shadows an α-spectrum casts when you lose the second axis. Primitive and environment in lambda calculus, data and procedure in the Lisp meta-circular interpreter, radial and angular time in `_archive/THEORY.md` — every one of these dualities is a pair of real projections of a single complex object, and the recurring error is mistaking the shadow for the thing.
 
 The coupled equation Z′ = α·Z + V·e^{iθ_v} does not balance two independent terms. It names one object — a coupling — whose two apparent components are each the other's source of incompressible novelty. Z is not a reflexive domain with V glued on; V is not an external signal entering a pre-existing Z. Each is the real projection of the same complex structure viewed from the other's axis. This is why the structural dependence theorem reads as a tautology once you see it — a coupling cannot exist with only one side, not because sides cooperate, but because "side" is already a projection artifact.
