+++ _archive/vybn_logic.md
