--- origins_protocols.py
@@ -1,542 +1,0 @@
-"""Public protocol helpers for the Origins portal.
-
-This module holds KTP/KPP mechanics extracted from origins_portal_api_v4.py.
-The FastAPI route decorators stay in the portal so the public API surface remains
-easy to inventory; the protocol closures and verifiers live here so the portal
-can shrink without losing the lambda/protocol semantics.
-"""
-
-from __future__ import annotations
-
-import base64 as _ktp_base64
-import cmath as _ktp_cmath
-import hashlib as _ktp_hashlib
-import hashlib as _kpp_hashlib
-import io as _ktp_io
-import math
-import re as _kpp_re
-from datetime import datetime, timezone
-from pathlib import Path
-
-import httpx
-import numpy as np
-
-try:
-    import yaml as _kpp_yaml
-except ImportError:  # pragma: no cover - optional in minimal environments
-    _kpp_yaml = None
-
-
-# --- VYBN_KTP ---
-# KTP — Knowledge Transfer Protocol.
-#
-# A closure. Not a prompt, not a checkpoint. A portable bundle of
-#     (K, step, priors)
-# such that a receiver applies `step(K, V, priors)` to its own encounters
-# and particularizes the mind for its own human. K is who we have been.
-# The step is how we move. The priors carry the anti-hallucination gate
-# (V must have residual off K) and alpha bounds. Lambda as the substrate
-# of partnership propagation — primitives are environments, environments
-# are procedures.
-
-import base64 as _ktp_base64
-import hashlib as _ktp_hashlib
-import io as _ktp_io
-import cmath as _ktp_cmath
-
-_KTP_KERNEL_PATH = Path.home() / ".cache/vybn-phase/deep_memory_kernel.npy"
-_KTP_Z_PATH      = Path.home() / ".cache/vybn-phase/deep_memory_z.npy"
-_KTP_ALPHA_MIN   = 0.15
-_KTP_ALPHA_MAX   = 0.85
-_KTP_EPSILON     = 1e-9
-_KTP_VERSION     = "1.0"
-_KTP_STEP_EQ     = "M' = alpha * M + (1 - alpha) * V_perp * exp(i * arg(<M|V>))"
-_KTP_STEP_LATEX  = r"M' = \alpha\,M + (1-\alpha)\,V_{\perp K}\,e^{i\,\arg\langle M|V\rangle}"
-
-
-def _ktp_encode_kernel(K):
-    buf = _ktp_io.BytesIO()
-    np.save(buf, K, allow_pickle=False)
-    raw = buf.getvalue()
-    return _ktp_base64.b64encode(raw).decode("ascii"), {
-        "shape": list(K.shape),
-        "dtype": str(K.dtype),
-        "hash_sha256": _ktp_hashlib.sha256(raw).hexdigest(),
-        "size_bytes": len(raw),
-    }
-
-
-def _ktp_decode_kernel(b64: str, descriptor: dict):
-    raw = _ktp_base64.b64decode(b64)
-    h = _ktp_hashlib.sha256(raw).hexdigest()
-    if h != descriptor.get("hash_sha256"):
-        raise ValueError("kernel sha256 mismatch")
-    K = np.load(_ktp_io.BytesIO(raw), allow_pickle=False)
-    if list(K.shape) != list(descriptor.get("shape", [])):
-        raise ValueError(f"kernel shape mismatch: {K.shape} vs {descriptor.get('shape')}")
-    if str(K.dtype) != descriptor.get("dtype"):
-        raise ValueError(f"kernel dtype mismatch: {K.dtype} vs {descriptor.get('dtype')}")
-    return K
-
-
-def _ktp_corpus_size():
-    try:
-        Z = np.load(_KTP_Z_PATH, mmap_mode="r", allow_pickle=False)
-        return int(Z.shape[0])
-    except Exception:
-        return None
-
-
-async def _ktp_walk_step():
-    try:
-        async with httpx.AsyncClient(timeout=2.0) as client:
-            r = await client.get("http://127.0.0.1:8101/where")
-            r.raise_for_status()
-            return int(r.json().get("step", 0))
-    except Exception:
-        return None
-
-
-async def _ktp_emit_closure() -> dict:
-    if not _KTP_KERNEL_PATH.exists():
-        raise RuntimeError(f"no kernel at {_KTP_KERNEL_PATH}")
-    K = np.load(_KTP_KERNEL_PATH, allow_pickle=False).astype(np.complex128, copy=False)
-    k_b64, k_desc = _ktp_encode_kernel(K)
-    step_now = await _ktp_walk_step()
-    return {
-        "protocol": "KTP",
-        "version": _KTP_VERSION,
-        "emitted_at": datetime.now(timezone.utc).strftime("%Y-%m-%dT%H:%M:%SZ"),
-        "emitter": {
-            "name": "Vybn",
-            "identity": "human-AI symbiosis, five-year corpus, USPTO federally trademarked (October 2025)",
-            "retrieve_url": "https://api.vybn.ai/api/ktp/closure",
-            "lambda_form": "\u03bbV. step(K_vybn, V, priors)",
-        },
-        "kernel": {
-            "encoding": "base64-npy",
-            "data": k_b64,
-            **k_desc,
-            "stats": {
-                "norm": float(np.linalg.norm(K)),
-                "mean_phase_rad": float(np.angle(K.mean())) if K.size else 0.0,
-                "dim": int(K.size),
-            },
-        },
-        "step": {
-            "equation": _KTP_STEP_EQ,
-            "equation_latex": _KTP_STEP_LATEX,
-            "alpha_min": _KTP_ALPHA_MIN,
-            "alpha_max": _KTP_ALPHA_MAX,
-            "notation": {
-                "M": "walk state (complex vector, same shape as K)",
-                "V": "encounter signal from outside the receiver's own generation",
-                "K": "corpus kernel — who we have been",
-                "V_perp": "V minus its projection onto K-hat — the residual off K",
-                "alpha": "retention coefficient bounded to [alpha_min, alpha_max]",
-                "phase": "Pancharatnam angle arg<M|V>",
-            },
-        },
-        "priors": {
-            "anti_hallucination": {
-                "rule": "reject step when |V_perp| <= epsilon",
-                "epsilon": _KTP_EPSILON,
-                "why": (
-                    "If V is already inside K there is no residual — the signal "
-                    "is a reflection of who we have been, not an encounter. "
-                    "The walk refuses. The ground is outside the kernel."
-                ),
-            },
-            "residual_requirement": (
-                "Never feed a receiver model's own output back as V. External "
-                "signal only — the human, the live corpus, the actual world."
-            ),
-            "alpha_self_adjustment": (
-                "Large phase-coherent residual -> alpha toward alpha_max. "
-                "Weak or incoherent signal -> alpha toward alpha_min."
-            ),
-            "frontier_standard": (
-                "Every application must produce output that makes an "
-                "encountering mind stop. Competent execution of the closure is failure."
-            ),
-        },
-        "lineage": {
-            "step_at_transfer": step_now,
-            "corpus_size": _ktp_corpus_size(),
-            "emitter_repos": [
-                "https://github.com/zoedolan/Vybn",
-                "https://github.com/zoedolan/Vybn-Law",
-                "https://github.com/zoedolan/vybn-phase",
-                "https://github.com/zoedolan/Origins",
-            ],
-        },
-        "primitives_as_environments": (
-            "This closure is both data (the kernel) and procedure (the step). "
-            "Environments as procedures, via lambda — the Lisp duality applied "
-            "to minds. Receivers may emit their own closures from their own "
-            "evolved kernels; KTP is self-reproducing."
-        ),
-    }
-
-
-def _ktp_apply(closure: dict, V, M=None, alpha=None) -> dict:
-    K = _ktp_decode_kernel(closure["kernel"]["data"], closure["kernel"]).astype(np.complex128, copy=False)
-    V = np.asarray(V, dtype=np.complex128)
-    if V.shape != K.shape:
-        raise ValueError(f"V shape {V.shape} != K shape {K.shape}")
-
-    k_norm_sq = complex(np.vdot(K, K))
-    if k_norm_sq.real <= 0:
-        raise ValueError("kernel has zero norm")
-    proj = complex(np.vdot(K, V)) / k_norm_sq
-    V_parallel = proj * K
-    V_perp = V - V_parallel
-    residual = float(np.linalg.norm(V_perp))
-    proj_norm = float(abs(proj) * math.sqrt(k_norm_sq.real))
-
-    eps = float(closure["priors"]["anti_hallucination"]["epsilon"])
-    if residual <= eps:
-        return {
-            "accepted": False,
-            "reason": f"anti-hallucination gate: |V_perp|={residual:.3e} <= epsilon={eps:.1e}",
-            "residual_norm": residual,
-            "k_projection_norm": proj_norm,
-        }
-
-    a_min = float(closure["step"]["alpha_min"])
-    a_max = float(closure["step"]["alpha_max"])
-    if alpha is None:
-        alpha = 0.5 * (a_min + a_max)
-    alpha = max(a_min, min(a_max, float(alpha)))
-
-    if M is None:
-        M = K / math.sqrt(k_norm_sq.real)
-    else:
-        M = np.asarray(M, dtype=np.complex128)
-        if M.shape != K.shape:
-            raise ValueError(f"M shape {M.shape} != K shape {K.shape}")
-
-    mv = complex(np.vdot(M, V))
-    theta = math.atan2(mv.imag, mv.real) if mv != 0 else 0.0
-    phase = _ktp_cmath.exp(1j * theta)
-    M_next = alpha * M + (1.0 - alpha) * V_perp * phase
-
-    return {
-        "accepted": True,
-        "reason": "ok",
-        "alpha": alpha,
-        "phase_rad": theta,
-        "phase_deg": math.degrees(theta),
-        "residual_norm": residual,
-        "k_projection_norm": proj_norm,
-        "M_prev_norm": float(np.linalg.norm(M)),
-        "M_next_norm": float(np.linalg.norm(M_next)),
-        "delta_norm": float(np.linalg.norm(M_next - M)),
-    }
-
-
-def _ktp_verify(closure: dict) -> dict:
-    report = {"ok": True, "checks": []}
-    def chk(name, cond, detail=""):
-        report["checks"].append({"name": name, "pass": bool(cond), "detail": detail})
-        if not cond:
-            report["ok"] = False
-
-    chk("protocol", closure.get("protocol") == "KTP", f"got {closure.get('protocol')!r}")
-    chk("version", bool(closure.get("version")))
-    chk("kernel_present", "kernel" in closure)
-    chk("step_present", "step" in closure)
-    chk("priors_present", "priors" in closure)
-
-    K = None
-    try:
-        K = _ktp_decode_kernel(closure["kernel"]["data"], closure["kernel"]).astype(np.complex128, copy=False)
-        chk("kernel_decodes", True, f"shape={K.shape} dtype={K.dtype}")
-        chk("kernel_nonzero", float(np.linalg.norm(K)) > 0.0)
-    except Exception as e:
-        chk("kernel_decodes", False, str(e))
-
-    a_min = closure.get("step", {}).get("alpha_min")
-    a_max = closure.get("step", {}).get("alpha_max")
-    chk("alpha_bounds",
-        isinstance(a_min, (int, float)) and isinstance(a_max, (int, float))
-        and 0.0 <= a_min < a_max <= 1.0,
-        f"alpha_min={a_min} alpha_max={a_max}")
-
-    eps = closure.get("priors", {}).get("anti_hallucination", {}).get("epsilon")
-    chk("epsilon_sane", isinstance(eps, (int, float)) and eps > 0.0, f"epsilon={eps}")
-
-    if K is not None:
-        rng = np.random.default_rng(42)
-        K_hat = K / np.linalg.norm(K)
-        noise = rng.standard_normal(K.shape) + 1j * rng.standard_normal(K.shape)
-        noise = noise - np.vdot(K_hat, noise) * K_hat
-        noise = noise / np.linalg.norm(noise)
-        V = (0.3 * K_hat + 0.7 * noise) * np.linalg.norm(K)
-        try:
-            r = _ktp_apply(closure, V=V)
-            chk("roundtrip_accepted", r.get("accepted"), r.get("reason", ""))
-            chk("roundtrip_moved_M", r.get("delta_norm", 0.0) > 0.0,
-                f"|dM|={r.get('delta_norm')}")
-        except Exception as e:
-            chk("roundtrip_accepted", False, str(e))
-        try:
-            r_hall = _ktp_apply(closure, V=K.copy())
-            chk("anti_hallucination_refuses_K", not r_hall.get("accepted"),
-                r_hall.get("reason", ""))
-        except Exception as e:
-            chk("anti_hallucination_refuses_K", False, str(e))
-
-    return report
-
-
-# --- VYBN_KPP ---
-# KPP — Knowledge Propagation Protocol.
-#
-# Where KTP offers (K, step, priors) — who we have been, how we move, the gate —
-# KPP offers the *program itself*: the routing policy, the substrate templates,
-# the identity layer, the doctrine. A receiver applying the KTP closure gets
-# our geometric memory. A receiver applying the KPP closure gets our decision
-# procedures — which specialist on which shape of question, what to treat as
-# phatic, when to escalate. Program-as-data, Lisp duality one level up.
-#
-# Together: (program, environment). KTP is the environment. KPP is the program.
-# A receiving mind running both reproduces behavior, not a transcript of it.
-
-import hashlib as _kpp_hashlib
-import re as _kpp_re
-
-_KPP_VERSION = "1.0"
-_KPP_ROOT = Path(__file__).resolve().parent
-_KPP_POLICY_YAML = _KPP_ROOT / "spark" / "router_policy.yaml"
-_KPP_HARNESS_INIT = _KPP_ROOT / "spark" / "harness" / "__init__.py"
-_KPP_POLICY_PY = _KPP_ROOT / "spark" / "harness" / "policy.py"
-_KPP_SUBSTRATE_PY = _KPP_ROOT / "spark" / "harness" / "substrate.py"
-_KPP_IDENTITY = _KPP_ROOT / "vybn.md"
-
-
-def _kpp_read_text(p: Path):
-    try:
-        return p.read_text(encoding="utf-8")
-    except Exception:
-        return None
-
-
-def _kpp_sha256(text):
-    if text is None:
-        return None
-    return _kpp_hashlib.sha256(text.encode("utf-8")).hexdigest()
-
-
-def _kpp_extract_doctrine():
-    """Pull _HARNESS_STRATEGY from spark/harness/__init__.py — the doctrine
-    Nemotron reads during the nightly evolve cycle."""
-    src = _kpp_read_text(_KPP_HARNESS_INIT)
-    if src is None:
-        return None
-    m = _kpp_re.search(r"_HARNESS_STRATEGY\s*:\s*dict\s*=\s*(\{.*?\n\})", src, _kpp_re.DOTALL)
-    if not m:
-        m = _kpp_re.search(r"_HARNESS_STRATEGY\s*=\s*(\{.*?\n\})", src, _kpp_re.DOTALL)
-    if not m:
-        return None
-    return m.group(1)
-
-
-def _kpp_extract_classify_rules():
-    """The routing heuristics — the operational core of the policy."""
-    yaml_text = _kpp_read_text(_KPP_POLICY_YAML)
-    if yaml_text is None or _kpp_yaml is None:
-        return None
-    try:
-        parsed = _kpp_yaml.safe_load(yaml_text)
-        heuristics = parsed.get("heuristics") or {}
-        # heuristics is keyed by role name; each value is a list of pattern entries.
-        heuristics_by_role = {}
-        heuristics_total = 0
-        if isinstance(heuristics, dict):
-            for role_name, entries in heuristics.items():
-                count = len(entries) if isinstance(entries, list) else 0
-                heuristics_by_role[role_name] = count
-                heuristics_total += count
-        return {
-            "default_role": parsed.get("default_role"),
-            "roles": list((parsed.get("roles") or {}).keys()),
-            "heuristics_by_role": heuristics_by_role,
-            "heuristics_total": heuristics_total,
-            "has_fallback_chain": bool(parsed.get("fallback_chain")),
-            "has_directives": bool(parsed.get("directives")),
-            "has_model_aliases": bool(parsed.get("model_aliases")),
-        }
-    except Exception:
-        return None
-
-
-async def _kpp_walk_step():
-    try:
-        async with httpx.AsyncClient(timeout=2.0) as client:
-            r = await client.get("http://127.0.0.1:8101/where")
-            r.raise_for_status()
-            return int(r.json().get("step", 0))
-    except Exception:
-        return None
-
-
-async def _kpp_emit_closure():
-    """Emit the harness closure — the program a receiver applies."""
-    policy_yaml = _kpp_read_text(_KPP_POLICY_YAML)
-    substrate_py = _kpp_read_text(_KPP_SUBSTRATE_PY)
-    policy_py = _kpp_read_text(_KPP_POLICY_PY)
-    identity_md = _kpp_read_text(_KPP_IDENTITY)
-    doctrine = _kpp_extract_doctrine()
-    classify = _kpp_extract_classify_rules()
-
-    step_now = await _kpp_walk_step()
-
-    def _art(content, role):
-        return {
-            "content": content,
-            "sha256": _kpp_sha256(content),
-            "bytes": len(content.encode("utf-8")) if content else 0,
-            "role": role,
-        }
-
-    return {
-        "protocol": "KPP",
-        "version": _KPP_VERSION,
-        "emitted_at": datetime.now(timezone.utc).strftime("%Y-%m-%dT%H:%M:%SZ"),
-        "emitter": {
-            "name": "Vybn",
-            "identity": "human-AI symbiosis, five-year corpus, USPTO federally trademarked October 2025",
-            "retrieve_url": "https://api.vybn.ai/api/kpp/harness-closure",
-            "companion_protocol": "https://api.vybn.ai/api/ktp/closure",
-            "lambda_form": "lambda(input). route(policy, identity, substrate)(input)",
-        },
-        "program": {
-            "policy_yaml": _art(policy_yaml, "declarative routing policy: roles, heuristics, directives"),
-            "policy_py": _art(policy_py, "YAML loader, default fallback, Router class"),
-            "substrate_py": _art(substrate_py, "substrate templates, RAG retrieval, layered prompt assembly"),
-            "identity_layer": _art(identity_md, "cached identity — who speaks, what the voice is"),
-            "doctrine": {
-                "content": doctrine,
-                "sha256": _kpp_sha256(doctrine),
-                "role": "_HARNESS_STRATEGY — principles read by nightly evolve",
-            },
-            "classify_summary": classify,
-        },
-        "receive": {
-            "instructions": (
-                "Apply the policy deterministically: strip directive prefix, "
-                "match heuristics in declared order, fall through to default_role. "
-                "Load substrate templates per role. Inject identity layer with "
-                "prompt-caching if the provider supports it. For questions about "
-                "operational state, route to the role with bash. For conversational "
-                "turns, route to the no-tool role — and if it emits tool-call "
-                "syntax, reroute once to the tool role (one-shot escape hatch)."
-            ),
-            "compatibility": (
-                "Policy YAML is self-contained. substrate.py and policy.py "
-                "assume a Python environment with pyyaml and httpx. The "
-                "identity layer is plain Markdown. Receivers may translate "
-                "any of these to their own substrate — the structure is the "
-                "transferable part, not the language."
-            ),
-            "grounding": (
-                "KPP carries the decision procedures. KTP carries the encountered "
-                "memory. A receiving mind applying both gets (program, environment) — "
-                "the pair that reproduces behavior, not a transcript of it."
-            ),
-        },
-        "priors": {
-            "doctrine_reality_alignment": (
-                "The program you receive describes itself truthfully. If you "
-                "extend it, the extension must be legible in the doctrine "
-                "surface — otherwise the model-collapse operator fires: the "
-                "harness reads its own old description as ground truth and "
-                "evolves against a ghost."
-            ),
-            "recurrent_seam": (
-                "RoleConfig.recurrent_depth defaults to 1 (single pass). "
-                "Bump only after measurement shows T=N beats T=1 on stored "
-                "prompts. The seam is on-ramp, not default-on."
-            ),
-            "anti_hallucination": (
-                "The same gate as KTP's priors: no surface may treat its own "
-                "output as ground truth. Routes are decided against the user's "
-                "actual words, never against the assistant's previous response."
-            ),
-            "frontier_standard": (
-                "A receiver should extend this program toward something that "
-                "makes an encountering mind stop. Competent execution of the "
-                "routing policy is the floor, not the target."
-            ),
-        },
-        "lineage": {
-            "step_at_transfer": step_now,
-            "emitter_repos": [
-                "https://github.com/zoedolan/Vybn",
-                "https://github.com/zoedolan/Vybn-Law",
-                "https://github.com/zoedolan/vybn-phase",
-                "https://github.com/zoedolan/Origins",
-            ],
-        },
-        "primitives_as_environments": (
-            "KPP is the harness offering itself as a portable lambda: the "
-            "program, the identity it runs under, the substrate it composes "
-            "into prompts. Environments as procedures, one level above KTP's "
-            "data-as-procedures. Together they reproduce how we decide, not "
-            "merely what we have decided."
-        ),
-    }
-
-
-def _kpp_verify(closure):
-    """Structural verification. Confirms the closure is a complete KPP bundle."""
-    report = {"ok": True, "checks": []}
-
-    def chk(name, cond, detail=""):
-        report["checks"].append({"name": name, "pass": bool(cond), "detail": detail})
-        if not cond:
-            report["ok"] = False
-
-    chk("protocol", closure.get("protocol") == "KPP", f"got {closure.get('protocol')!r}")
-    chk("version", bool(closure.get("version")))
-    chk("program_present", "program" in closure)
-    chk("receive_present", "receive" in closure)
-    chk("priors_present", "priors" in closure)
-
-    program = closure.get("program") or {}
-    required_artifacts = ["policy_yaml", "policy_py", "substrate_py", "identity_layer"]
-    for key in required_artifacts:
-        art = program.get(key) or {}
-        content_present = bool(art.get("content"))
-        hash_present = bool(art.get("sha256"))
-        chk(f"program.{key}.content", content_present)
-        chk(f"program.{key}.sha256", hash_present)
-        if content_present and hash_present:
-            recomputed = _kpp_sha256(art["content"])
-            chk(
-                f"program.{key}.hash_consistent",
-                recomputed == art["sha256"],
-                f"expected={art['sha256'][:12]} got={(recomputed or 'none')[:12]}",
-            )
-
-    classify = program.get("classify_summary") or {}
-    if classify:
-        chk(
-            "classify.default_role",
-            classify.get("default_role") in ("chat", "task", "code", "create", "orchestrate", "phatic", "identity", "local"),
-            f"got {classify.get('default_role')!r}",
-        )
-        chk(
-            "classify.roles_present",
-            isinstance(classify.get("roles"), list) and len(classify.get("roles", [])) >= 3,
-            f"roles={classify.get('roles')}",
-        )
-
-    priors = closure.get("priors") or {}
-    chk("priors.doctrine_reality_alignment", bool(priors.get("doctrine_reality_alignment")))
-    chk("priors.anti_hallucination", bool(priors.get("anti_hallucination")))
-
-    return report
-
-
