+++ spark/tests/test_harness.py
@@ -1748,7 +1748,7 @@ class CommonsWalkTests(unittest.TestCase):
             self.assertTrue(manifest["entrypoints"], name)
             self.assertTrue(manifest["agentActions"], name)
             self.assertTrue(manifest["traceProtocol"], name)
-            self.assertEqual(manifest["ontology"], "https://raw.githubusercontent.com/zoedolan/Vybn/main/commons-skeleton.json")
+            self.assertEqual(manifest["ontology"], "https://raw.githubusercontent.com/zoedolan/Vybn/main/_archive/commons-skeleton.json")
             self.assertEqual(manifest["encounterLifecycle"], skeleton["encounterLifecycle"])
             self.assertEqual(manifest["aiNativePrinciple"], AI_NATIVE_PRINCIPLE)
             self.assertTrue(manifest["dynamicAffordanceProtocol"], name)
